restructure plugin files
